requirements.txt
# Core AI/ML dependencies
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4

# LLaMA integration and quantization
huggingface-hub>=0.19.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Ollama integration (alternative to Hugging Face)
ollama>=0.1.7

# Data processing and manipulation
pandas>=2.0.0
numpy>=1.24.0
PyYAML>=6.0

# Web interface and UI
streamlit>=1.29.0
streamlit-chat>=0.1.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0

# Progress bars and utilities
tqdm>=4.66.0
python-dotenv>=1.0.0

# JSON and text processing
jsonschema>=4.17.0

# Logging and system
colorlog>=6.7.0
psutil>=5.9.0

# HTTP requests
requests>=2.31.0

# Optional: GPU support (uncomment if you have NVIDIA GPU)
# faiss-gpu>=1.7.4  # Replace faiss-cpu with this for GPU support

# Optional: Additional ML utilities (install if needed)
# scikit-learn>=1.3.0
# matplotlib>=3.7.0
# seaborn>=0.12.0

# Optional: Development tools (install if needed)
# pytest>=7.4.0
# pytest-cov>=4.1.0
# black>=23.7.0
# flake8>=6.0.0

# Optional: Text processing enhancements (install if needed)
# nltk>=3.8.0
# spacy>=3.6.0

# Optional: Additional Streamlit components (install if needed)
# streamlit-option-menu>=0.3.6
# streamlit-aggrid>=0.3.4

# Optional: File handling (install if needed)
# openpyxl>=3.1.0
# xlsxwriter>=3.1.0
